consolidated DP matrix implementations to one file
