+++ libnail/src/align/structs/mod.rs
@@ -11,13 +11,7 @@ mod cloud_search_params;
 pub use cloud_search_params::CloudSearchParams;
 
 mod dp_matrix;
-pub use dp_matrix::DpMatrix;
-
-mod dp_matrix_flat;
-pub use dp_matrix_flat::DpMatrixFlat;
-
-mod dp_matrix_sparse;
-pub use dp_matrix_sparse::DpMatrixSparse;
+pub use dp_matrix::{DpMatrix, DpMatrixFlat, DpMatrixSparse};
 
 mod row_bounds;
 pub use row_bounds::RowBounds;
